+++ nerdlets/slo-r-main/components/slo-summary/index.js
@@ -1,5 +1,5 @@
 /**
- * Provides the component that a rolled up SLO attainment for an Organization
+ * Provides the component that a rolled up SLO attainment for a SLO Group
  *
  * @file
  * @author Gil Rice
@@ -9,7 +9,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 /** nr1 */
-import { HeadingText, Spinner, Stack, StackItem } from 'nr1';
+import {
+  Button,
+  HeadingText,
+  navigation,
+  Spinner,
+  Stack,
+  StackItem
+} from 'nr1';
 
 /** local */
 import CompositeAlertSlo from '../../../shared/queries/alert-driven-slo/composite';
@@ -20,11 +27,11 @@ import { SLO_INDICATORS } from '../../../shared/constants';
 import BootstrapTable from 'react-bootstrap-table-next';
 
 /**
- * OrgDisplayer
+ * SloSummary
  */
-export default class OrgDisplayer extends React.Component {
+export default class SloSummary extends React.Component {
   static propTypes = {
-    org: PropTypes.object,
+    tag: PropTypes.object,
     timeRange: PropTypes.object,
     activeIndicator: PropTypes.string
   }; // propTypes
@@ -42,21 +49,21 @@ export default class OrgDisplayer extends React.Component {
 
   async componentDidMount() {
     this.filterSlos();
-    await this._assembleOrganizationData();
+    await this._assembleData();
   } // componentDidMount
 
   async componentDidUpdate(prevProps) {
-    const orgChanged = prevProps.org.orgName !== this.props.org.orgName;
+    const tagChanged = prevProps.tag.tagName !== this.props.tag.tagName;
     const timeRangeChanged = prevProps.timeRange !== this.props.timeRange;
     const selectedIndicatorChanged =
       prevProps.activeIndicator !== this.props.activeIndicator;
 
-    if (orgChanged || timeRangeChanged || selectedIndicatorChanged) {
-      await this._assembleOrganizationData();
+    if (tagChanged || timeRangeChanged || selectedIndicatorChanged) {
+      await this._assembleData();
     }
   }
 
-  async _assembleOrganizationData() {
+  async _assembleData() {
     const { activeIndicator } = this.props;
     const slosFilteredByIndicator = this.filterSlos();
 
@@ -98,7 +105,7 @@ export default class OrgDisplayer extends React.Component {
       summarySloData,
       loadingData: false
     });
-  } // _assembleOrganizationData
+  } // _assembleData
 
   transformToTableData({ data }) {
     const { activeIndicator } = this.props;
@@ -115,6 +122,9 @@ export default class OrgDisplayer extends React.Component {
   /* Transform row data for bootstrap table */
   transformRow({ data }) {
     const transformedData = {
+      documentId: data.slo_document.documentId,
+      entityGuid: data.slo_document.entityGuid,
+
       name: data.slo_document.name,
       target: data.slo_document.target,
 
@@ -127,8 +137,8 @@ export default class OrgDisplayer extends React.Component {
   }
 
   filterSlos() {
-    const { activeIndicator, org } = this.props;
-    const slosFilteredByIndicator = org.slos.filter(
+    const { activeIndicator, tag } = this.props;
+    const slosFilteredByIndicator = tag.slos.filter(
       item => item.slo.indicator === activeIndicator
     );
 
@@ -187,7 +197,7 @@ export default class OrgDisplayer extends React.Component {
   }
 
   renderBootStrapTableView() {
-    const { activeIndicator, org } = this.props;
+    const { activeIndicator, tag } = this.props;
     const { tableData, summarySloData } = this.state;
     const attainment = this.calculateTotalAttainment({
       _slo_data: summarySloData
@@ -195,11 +205,35 @@ export default class OrgDisplayer extends React.Component {
 
     const indicatorLabel = SLO_INDICATORS.find(i => i.value === activeIndicator)
       .label;
-    const tableHeader = `${org.orgName}'s ${indicatorLabel} SLO's`;
+    const tableHeader = `${tag.tagName}'s ${indicatorLabel} SLO's`;
+
+    // eslint-disable-next-line no-unused-vars
+    const linkFormatter = function(cell, row, rowIndex, formatExtraData) {
+      const { entityGuid } = row;
+      return (
+        <Button
+          type={Button.TYPE.NORMAL}
+          iconType={Button.ICON_TYPE.INTERFACE__OPERATIONS__EXTERNAL_LINK}
+          onClick={() => {
+            navigation.openStackedEntity(entityGuid);
+          }}
+        />
+      );
+    };
 
     const columns = [
       {
-        dataField: 'name', // SLO
+        dataField: 'entityGuid',
+        text: 'Entity Guid',
+        hidden: true
+      },
+      {
+        dataField: 'documentId',
+        text: 'SLO Document Id',
+        hidden: true
+      },
+      {
+        dataField: 'name', // SLO Name not Entity Name
         text: 'Name',
         footer: 'Total attainment:',
         headerStyle: () => {
@@ -225,6 +259,12 @@ export default class OrgDisplayer extends React.Component {
         dataField: 'target',
         text: 'Target',
         footer: '--'
+      },
+      {
+        dataField: '',
+        text: 'Entity',
+        footer: '--',
+        formatter: linkFormatter
       }
     ];
 
@@ -296,5 +336,5 @@ export default class OrgDisplayer extends React.Component {
     }
 
     return <>{this.renderBootStrapTableView()}</>;
-  } // render
-} // OrgDisplayer
+  }
+}